rename projects to pixi_rocker
