+++ .devcontainer/devcontainer.json
@@ -1,5 +1,5 @@
 {
-	"name": "python_template",
+	"name": "pixi_rocker",
 	"build": {
 		"dockerfile": "Dockerfile",
 		"context": ".."
